+++ src/KOKKOS/compute_surf_kokkos.h
@@ -445,7 +445,7 @@ void surf_tally_kk(double /*dtremain*/, int isurf, int icell, int reaction,
 }
 
  private:
-  int mvv2e;
+  double mvv2e;
   int compressed;     // 1 once the device tallies have been copied and compressed
 
   DAT::t_int_1d d_which;
